Update for new array/addr indexing primops.
The simd branch of GHC now has primops for indexing into ByteArray's and Addr's
using offsets expressed in terms of type a but reading/writing values of type
Multi a. This patch adds class members to the MultiPrim class that makes using
these primops easier. All instances of Multi and MultiPrim are also now defined
in Data.Primitive.Multi.